#27 - moves monthly limit component to home module
