+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/HomeScreen.kt
@@ -44,9 +44,9 @@ import br.com.rbrthmn.R
 import br.com.rbrthmn.home.ui.components.balancecard.BalanceCard
 import br.com.rbrthmn.home.ui.components.creditcardbillscard.CreditCardBillsCard
 import br.com.rbrthmn.home.ui.components.lastmonthdifferencecard.LastMonthDifferenceCard
+import br.com.rbrthmn.home.ui.components.monthlylimitcard.MonthlyLimitCard
 import br.com.rbrthmn.navigation.NavigationDestination
 import br.com.rbrthmn.ui.components.MonthSelectionTopBar
-import br.com.rbrthmn.ui.financialcompanion.screens.home.components.monthlylimitcard.MonthlyLimitCard
 import br.com.rbrthmn.ui.financialcompanion.utils.SnackBarProvider
 import org.koin.androidx.compose.koinViewModel
 import org.koin.compose.koinInject
